# backend/openAPI.yaml
...
openapi: 3.0.3
info:
  title: Knowledge Graph API
  version: 1.1.0
  description: |
    Knowledge Graph API with unified REST responses.
    All responses follow the standard SuccessResponse and ErrorResponse formats.

    ## Authentication

    Authenticated endpoints accept either:
    - `Authorization: Bearer <JWT>` header, or
    - HttpOnly `access_token` cookie (sent automatically by the browser).

    ## Response format

    ### Success (200, 201):
    ```json
    {
      "data": { ... },
      "message": "Optional success message"
    }
    ```

    ### Error (400, 401, 403, 404, 409, 500):
    ```json
    {
      "code": "ERROR_CODE",
      "message": "Human-readable description",
      "details": [
        {
          "field": "field",
          "reason": "reason",
          "message": "description",
          "received": "received value",
          "expected": "expected value"
        }
      ]
    }
    ```
servers:
  - url: http://127.0.0.1:18080
    description: Dev nginx gateway (host port)
  - url: http://127.0.0.1:18083
    description: Test backend direct (SKIP_AUTH=false test stack)
  - url: http://127.0.0.1:9000
    description: Dev backend direct
paths:
...
  /api/v1/notes/{id}:
    get:
      summary: Get note by ID
      tags: [Notes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Note found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      summary: Update note
      tags: [Notes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNoteRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      summary: Delete note
      tags: [Notes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        data:
          type: object
          description: Response data (note, link, graph, etc.)
        message:
          type: string
          description: Optional success message
      example:
        data:
          id: "550e8400-e29b-41d4-a716-446655440000"
          title: "Note Title"
          content: "Note content"
        message: "Resource created successfully"
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum: [VALIDATION_ERROR, NOT_FOUND, CONFLICT, FORBIDDEN, UNAUTHORIZED, INTERNAL_ERROR]
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
      example:
        code: "VALIDATION_ERROR"
        message: "Invalid input data"
        details:
          - field: "title"
            reason: "required"
            message: "Title is required"
    FieldError:
      type: object
      required:
        - field
        - reason
        - message
      properties:
        field:
          type: string
          description: Field name with error
        reason:
          type: string
          description: Error reason code
          enum: [required, invalid_type, invalid_format, already_exists, too_long, out_of_range, invalid_value]
        message:
          type: string
          description: Human-readable error description
        received:
          type: string
          description: Received value (optional)
        expected:
          type: string
          description: Expected value/format (optional)
...
    UpdateNoteRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        content:
          type: string
          maxLength: 50000
        metadata:
          type: object
...
  responses:
    BadRequest:
      description: Bad Request - Validation Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "VALIDATION_ERROR"
            message: "Invalid input data"
            details:
              - field: "title"
                reason: "required"
                message: "Title is required"
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "NOT_FOUND"
            message: "Note not found"
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "FORBIDDEN"
            message: "Access denied"
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "INTERNAL_ERROR"
            message: "Internal server error"